+++ test_unstructured/partition/pdf_image/test_pdf.py
@@ -969,3 +969,10 @@ def test_partition_pdf_with_ocr_only_strategy(
     # check detection origin
     if UNSTRUCTURED_INCLUDE_DEBUG_METADATA:
         assert {element.metadata.detection_origin for element in elements} == {"ocr_tesseract"}
+
+
+def test_partition_pdf_with_all_number_table_and_ocr_only_strategy():
+    # AttributeError was previously being raised when partitioning documents that contained only
+    # numerical values with `strategy="ocr_only"`
+    filename = example_doc_path("all-number-table.pdf")
+    assert pdf.partition_pdf(filename, strategy="ocr_only")