+++ unstructured/partition/ocr.py
@@ -533,7 +533,9 @@ def parse_ocr_data_tesseract(ocr_data: pd.DataFrame, zoom: float = 1) -> List[Te
         text = idtx.text
         if not text:
             continue
-        cleaned_text = text.strip()
+
+        cleaned_text = str(text) if not isinstance(text, str) else text.strip()
+
         if cleaned_text:
             x1 = idtx.left / zoom
             y1 = idtx.top / zoom